+++ packages/envsync-enterprise/src/routes/license.route.ts
@@ -2,7 +2,7 @@ import { Hono } from "hono";
 import { describeRoute } from "hono-openapi";
 import { resolver } from "hono-openapi/zod";
 
-import { LicenseController } from "@/controllers/license.controller";
+import { LicenseController } from "../controllers/license.controller";
 import { errorResponseSchema } from "@/validators/common";
 import { licenseActionResponseSchema, licenseStatusResponseSchema } from "@/validators/license.validator";