+++ packages/envsync-enterprise/src/routes/oidc.route.ts
@@ -2,7 +2,7 @@ import { Hono } from "hono";
 import { describeRoute } from "hono-openapi";
 import { resolver, validator as zValidator } from "hono-openapi/zod";
 
-import { OidcController } from "@/controllers/oidc.controller";
+import { OidcController } from "../controllers/oidc.controller";
 import { authMiddleware } from "@/middlewares/auth.middleware";
 import { enterpriseGuard } from "@/middlewares/enterprise.middleware";
 import { requirePermission } from "@/middlewares/permission.middleware";
@@ -11,7 +11,7 @@ import {
 	oidcProviderResponseSchema,
 	oidcProvidersResponseSchema,
 	updateOidcProviderRequestSchema,
-} from "@/validators/oidc.validator";
+} from "../validators/oidc.validator";
 import { errorResponseSchema } from "@/validators/common";
 
 const app = new Hono();